+++ examples/expand.py
@@ -6,7 +6,12 @@ This script demonstrates using the HeritagedataProvider to expand a concept
 
 from skosprovider_heritagedata.providers import HeritagedataProvider
 
-results = HeritagedataProvider({'id': 'Heritagedata'}, scheme_uri='http://purl.org/heritagedata/schemes/eh_period').expand('PM')
+periodprovider = HeritagedataProvider(
+    {'id': 'Heritagedata'},
+    scheme_uri='http://purl.org/heritagedata/schemes/eh_period'
+)
+
+results = periodprovider.expand('PM')
 
 print('Results')
 print('------')