+++ examples/find.py
@@ -6,10 +6,19 @@ This script demonstrates using the HeritagedataProvider to find the concepts wit
 
 from skosprovider_heritagedata.providers import HeritagedataProvider
 
-results = HeritagedataProvider({'id': 'Heritagedata'}, scheme_uri='http://purl.org/heritagedata/schemes/eh_period').find({'label': 'iron', 'type': 'concept'})
+periodprovider = HeritagedataProvider(
+    {'id': 'Heritagedata'},
+    scheme_uri='http://purl.org/heritagedata/schemes/eh_period'
+)
+
+results = periodprovider.find(
+    {
+        'label': 'iron',
+        'type': 'concept'
+    }
+)
 
 print('Results')
 print('------')
 for result in results:
     print(result)
-