+++ CHANGES.rst
@@ -1,9 +1,14 @@
-0.2.0 (2014-??-??)
+0.2.0 (2014-12-??)
 ------------------
 
-- Compatibile with `SkosProvider 0.4.0 <http://skosprovider.readthedocs.org/en/0.4.0>`_.
+- Compatibile with `SkosProvider 0.5.0 <http://skosprovider.readthedocs.org/en/0.4.0>`_.
 - Not backwards compatible with `SkosProvider 0.3.0 <http://skosprovider.readthedocs.org/en/0.3.0>`_.
+- Each Concept or Collection now also provides information on the ConceptScheme 
+  it's part of.
 - Fix some issues with UTF-8 encoding.
+- Fixed some issues with Python 2.x/3.x compatibility.
+- Provider now throws a ProviderNotAvailableException when the underlying service is down.
+
 
 0.1.0 (2014-10-08)
 ------------------